// packages/dsh-aionui-panel/src/core/types.ts
export type PreviewContentType =
  | 'markdown'
  | 'html'
  | 'code'
  | 'diff'
  | 'csv'
  | 'pdf'
  | 'word'
  | 'excel'
  | 'ppt'
  | 'image'
  | 'text'
  | 'url'
  | 'unsupported'